# src/main/resources/openapi/pricelist-api.yaml
---
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  "/v1/price-lists":
    get:
      tags:
        - price-list-controller
      summary: Get price-tariff applied to the product within a date-range
      description: Find the price-tariff based on a productId, brandId and effectiveDate
      operationId: getPriceListByDateProductBrand
      parameters:
        - name: effective-date
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: product-id
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: brand-id
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PriceListResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponseDetails"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponseDetails"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponseDetails"
components:
  schemas:
    ErrorResponseDetails:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
    PriceListResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        brandId:
          type: integer
          format: int32
        priceListId:
          type: integer
          format: int64
        price:
          type: number
        currency:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time